move all the svg files to assests folder
